# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Picking-software API - Auth and Estimates",
    "version": "1.0.0",
    "description": "API specifications for authentication and estimate-related endpoints"
  },
  "servers": [
    {
      "url": "https://api.smartpicker.au",
      "variables": {
        "port": {
          "default": "443"
        }
      }
    }
  ],
  "tags": [
    {
      "name": "User Auth",
      "description": "Authentication endpoints"
    },
    {
      "name": "Estimates",
      "description": "Estimate-related endpoints"
    }
  ],
  "paths": {
    "/authUri": {
      "get": {
        "summary": "Get authentication URI",
        "tags": ["User Auth"],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/callback": {
      "get": {
        "summary": "OAuth callback",
        "tags": ["User Auth"],
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to frontend"
          }
        }
      }
    },
    "/csrf-token": {
      "get": {
        "summary": "Get CSRF token",
        "tags": ["User Auth"],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "csrfToken": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/verifyUser": {
      "get": {
        "summary": "Verify user",
        "tags": ["User Auth"],
        "responses": {
          "200": {
            "description": "User is valid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "isValid": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "User is not valid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "isValid": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "csrfToken": []
          }
        ]
      }
    },
    "/getEstimates/{customerId}": {
      "get": {
        "summary": "Get estimates for a customer",
        "tags": ["Estimates"],
        "parameters": [
          {
            "in": "path",
            "name": "customerId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "csrfToken": []
          }
        ]
      }
    },
    "/estimate/{quoteId}": {
      "get": {
        "summary": "Get estimate by ID",
        "tags": ["Estimates"],
        "parameters": [
          {
            "in": "path",
            "name": "quoteId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "csrfToken": []
          }
        ]
      }
    },
    "/quotes": {
      "get": {
        "summary": "Get quotes by status",
        "tags": ["Estimates"],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "csrfToken": []
          }
        ]
      }
    },
    "/quote-status": {
      "put": {
        "summary": "Update quote status",
        "tags": ["Estimates"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quoteId": {
                    "type": "string"
                  },
                  "newStatus": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "csrfToken": []
          }
        ]
      }
    },
    "/updateQuoteInQuickBooks/{quoteId}": {
      "put": {
        "summary": "Update quote in QuickBooks",
        "tags": ["Estimates"],
        "parameters": [
          {
            "in": "path",
            "name": "quoteId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "csrfToken": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    }
  }
}